ripapay/backend/requirements.txt
qubipy==0.2.5
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
qrcode>=7.3
pillow>=9.0.0  # Required for QR code image generation